Obtaining the previous frame stack pointer
When running the "DW_OP_call_frame_cfa" opcode, LLDB set it to
"Value::eValueTypeLoadAddress" type, which means the value is an address
that must be read to obtain the correct CFA value.
The issue is that when trying to read this value, LLDB fails as the compiler
type was not set.
In order to fix it, the memory read was added just before the address be
obtained.
This issue as making gcc binaries print wrong variable values, as it could
not find the correct address of the values.